# Split StellarGraph benchmarks into creation and adj list ones